requirements.txt
streamlit==1.29.0
pandas==2.1.3
openpyxl==3.1.2
pillow==10.1.0